kmu-define-keys-1: don't remove new bindings again
If a key is bound to some command and later it is removed again, then
ignore the second instruction.  Now if we want to say both 'bind "b" to
what "a" was previously bound to', as well as 'bind "c" to what "b" was
previously bound to', we can just say:

  (kmu-define-key some-map nil
    "a" > "b" command-1
    "b" > "c" command-2)

Previously that would have failed because '"b" >' removes the binding
established with '"b" command-1.

Instead we would have had to say:

  (kmu-define-key some-map nil
    "a" > "b" command-1
    "b" ~ "c" command-2)

But if these two lines were far apart, then it would be easy to not
realize that this was necessary.

Also remove support for this `~' no-op.

Note that the two keys have to be `equal' for the new approach to work.