kmu-define-keys: expand to nicer form
Expand to a call to `kmu-remove-key' or `kmu-define-key', or a `progn'
or `eval-after-load' containing such forms, instead of something gross
which does the work that is now done at macro-expansion time.

This is done mainly so that users of e.g. `macrostep-expand' have
something pretty and useful to look at, not for performance reasons.

Remove the now unnecessary `kmu-define-keys-1' helper.  In the progress
we lose a sanity check ("is the keymap really a keymap?") and it is no
longer possible to abuse the helper to be able to directly modify a
keymap instead of the keymap that is the value of a variable.