kmu-remove-key: add kludge for global-map
In the `global-map' meta is bound to the symbol `ESC-prefix' and
build-in functions that work with keymaps seem to deal with that
just fine.  Add a special case to `kmu-remove-key' so that this
holds true for that function too.

Also remove the function `kmu-merge-esc-into-global-map' since
`kmu-remove-key' was the only function that required that "fix".