kmu-define-key: Ensure parent keymaps aren't modified
Given
    (keymap
     (97 . undesired)
     keymap
     (97))
by removing "a" we got
    (keymap
     keymap)
now we get the expected
    (keymap
     keymap
     (97))

Remove only the first appearance of (EVENT) in keymap, whether it is
the top-level keymap or a submap.  For full keymaps do not attempt to
remove (EVENT), which does not exist when EVENT's binding is stored in
the char-table.