Use same key representation as Emacs 29.1
When this package was created, Emacs represented keys differently
in code and documentation.  One had to use strings like "\C-c\" or
vectors like [\?C-c], and in documentation that appeared as "C-c".

Emacs 29.1 adds new keymap functions that use the human friendly
representation we are familiar with from documentation.  For the most
part this package already did that in the past, but unfortunately it
diverted somewhat.  Taking inspiration from "naked.el" [1], it did not
require angle brackets around named events, so you could, write just
"return" instead of "<return>".

That seemed like a good idea at the time (but now I find it highly
questionable), given that the keymap functions did not support the
human friendly format anyway, with or without angle brackets.  But
now that there is a new set of functions that do, it is important to
support exactly the same format here.  The vector representation also
isn't supported anymore.

Luckily the byte-compiler will helpfully point out any keys that are
invalid after this breaking change, i.e., those that are not
`key-valid-p'.

Also remove functions that are no longer necessary because the new
official keymap functions can be used instead: `kmu-key-description',
`kmu-parse-key-description', `kmu-define-key', and `kmu-remove-key'
(use `keymap-unset' with non-nil REMOVE argument).

[1] https://www.emacswiki.org/emacs/NaKeD